+++ attributes/default.rb
@@ -1,4 +1,3 @@
-default["cida_rabbitmq"]["credentials_data_bag_name"] = "cida_rabbitmq-_default"
-default["cida_rabbitmq"]["credentials_data_bag_item"] = "cida_rabbitmq_credentials"
-default["cida_rabbitmq"]["data_bag_username_field"] = "username"
-default["cida_rabbitmq"]["data_bag_password_field"] = "password"
+default['cida_rabbitmq']['credentials_data_bag_name'] = 'cida_rabbitmq-_default'
+default['cida_rabbitmq']['credentials_data_bag_item'] = 'cida_rabbitmq_credentials'
+default['cida_rabbitmq']['data_bag_users_field'] = 'users'