+++ metadata.rb
@@ -1,10 +1,13 @@
 name             'cida_rabbitmq'
 maintainer       'Ivan Suftin'
 maintainer_email 'isuftin@usgs.gov'
-license          'Public Domain'
+license          'CPL-1.0'
 description      'Installs/Configures Rabbit MQ with sane defaults.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.5'
-
-depends 'iptables'
+version          '1.0.0'
+chef_version '>= 12.0.0'
+supports         'centos', '>= 6.6'
+supports         'centos', '>= 7.1'
 depends 'rabbitmq'
+source_url 'https://github.com/USGS-CIDA/cookbook-rabbitmq'
+issues_url 'https://github.com/USGS-CIDA/cookbook-rabbitmq/issues'