+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Changelog
 ---------
 
+- 1.0.0
+  -----
+  - [isuftin@usgs.gov] - Working with latest RabbitMQ cookbook
+  - [isuftin@usgs.gov] - Changed default recipe to do nothing
+  - [isuftin@usgs.gov] - Removed iptables recipe
+  - [isuftin@usgs.gov] - Create users now wraps the upstream cookbook recipe
+
 - 0.1.5
   -----
   - [isuftin@usgs.gov] - Fixed a kitchen.yml issue that broke testing
@@ -10,9 +17,9 @@
 
 - 0.1.4
   -----
-  - [isuftin@usgs.gov] - Refactored .kitchen.yml to include more images, generalize ports and attributes, etc 
+  - [isuftin@usgs.gov] - Refactored .kitchen.yml to include more images, generalize ports and attributes, etc
   - [isuftin@usgs.gov] - Added integration testing
   - [isuftin@usgs.gov] - Broke up default recipe
   - [isuftin@usgs.gov] - Updated default attributes and removed the need to specify an encryption key location
   - [isuftin@usgs.gov] - Removed 3rd party RabbitMQ cookbook attributes and recipe calls out of cookbook and into
-  						.kitchen.yml 
+  						.kitchen.yml